requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
pydantic==2.5.0
sqlalchemy==2.0.23
requests==2.31.0
python-dotenv==1.0.0
